Checked CTA.getLabelSetID and extended time measurement in CTA-intersection
Fixed:	Changing content of set, that is used as a key in a hashmap in CondensedTreeAutomaton
Fixed:	CTA now seperates between adding a labelSet to an Interner and just getting its ID.
Added:	Extended Interner for IntSets, that helps updating the content of IntSets that are used in an Interner.

Added:	Time measurement in CTA-Intersection (main-class) is now based on CPU time, if available.
Added:	Triggering garbage collection after each sentence in CTA-Intersection