Interpretation: Generics, reduce string-based lookups
This commits adds type safety to some parts dealing with
interpretations.  Methods now take an Interpretation directly instead
of the name of the interpretation.  This a) reduces lookups and b)
provides type safety as the return types of these methods now
correspond to the interpretation being used.

An interpretation now knows its own name.